FM-APP: Fix tune failed error when headset is plugged in/out
While repeatedly plugging in/out the headset, we observe
the message "FM operation failed. If problem persists,
please turn off FM and turn it back on" on the UI.

Just after disabling fm, call stop in order to ensure that
fm is in proper state.
Before calling tune make sure to check whether FM is
in proper state.

Change-Id: I718652d17849d869a48d7dc3ca608adf8ab51ce8
CRs-Fixed: 331107